pyproject.toml
[project]
name = "locawise"
version = "1.0.0-beta"
description = ""
authors = [
    {name = "Emre Safak",email = "aemresafak@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jproperties (>=2.1.2,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "xxhash (>=3.5.0,<4.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pycountry (>=24.6.1,<25.0.0)",
    "google-genai (>=1.10.0,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "openai (>=1.76.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "locawise", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "module"
asyncio_default_fixture_loop_scope = "module"

